+++ source/common/zettlr-helpers.js
@@ -373,6 +373,29 @@ function localiseNumber(number)
     return ret;
 }
 
+/**
+ * This function takes a Markdown string and replaces all occurrences of images
+ * with an absolutised version.
+ * @param  {String} basepath The basepath with which relative paths should be joined.
+ * @param  {String} mdstring The string to be altered.
+ * @return {String}          The altered mdstring value.
+ */
+function makeImgPathsAbsolute(basePath, mdstring)
+{
+    let imgRE = /^!\[(.*?)\]\((.+?)\)$/gmi;
+    let match;
+    return mdstring.replace(imgRE, (match, p1, p2, offset, string) => {
+        // Check if the path (p2) contains the absolute path
+        if(p2.indexOf(basePath) === 0 || p2.indexOf('http') === 0) {
+            // It's already absolute (either local or remote)
+            return `![${p1}](${p2})`;
+        } else {
+            // Make it absolute
+            return `![${p1}](${path.join(basePath, p2)})`;
+        }
+    });
+}
+
 module.exports = {
     hash,
     flattenDirectoryTree,
@@ -386,5 +409,6 @@ module.exports = {
     isDir,
     isAttachment,
     localiseNumber,
-    isDictAvailable
+    isDictAvailable,
+    makeImgPathsAbsolute
 };